requirements.txt
Django==4.2.3
django-crispy-forms==2.0
django-filter==23.2
Pillow==10.0.0
crispy-bootstrap4-2022.1